+++ package.json
@@ -67,7 +67,7 @@
     "react-dom": "18.2.0",
     "tsm": "^2.2.2",
     "tsup": "^6.2.3",
-    "typescript": "^4.8.2",
+    "typescript": "^4.9.5",
     "vi-fetch": "^0.8.0",
     "vite": "^4.0.4",
     "vitest": "^0.28.1"
