+++ src/handlers/next/index.ts
@@ -2,4 +2,5 @@ export * from "./create-app-register-handler";
 export * from "./create-manifest-handler";
 export * from "./create-protected-handler";
 export * from "./process-protected-handler";
-export * from "./saleor-async-webhook";
+export * from "./saleor-webhooks/saleor-async-webhook";
+export * from "./saleor-webhooks/saleor-sync-webhook";
