+++ src/handlers/next/saleor-webhooks/saleor-async-webhook.test.ts
@@ -2,7 +2,7 @@ import { ASTNode } from "graphql";
 import { createMocks } from "node-mocks-http";
 import { afterEach, describe, expect, it, vi } from "vitest";
 
-import { APL } from "../../APL";
+import { APL } from "../../../APL";
 import { processSaleorWebhook } from "./process-saleor-webhook";
 import { NextWebhookApiHandler, SaleorAsyncWebhook } from "./saleor-async-webhook";
