+++ src/handlers/next/saleor-webhooks/saleor-sync-webhook.ts
@@ -1,0 +1,14 @@
+import { SyncWebhookEventType } from "../../../types";
+import { SaleorWebhook, WebhookConfig } from "./saleor-webhook";
+
+export class SyncSaleorWebhook<TPayload = unknown> extends SaleorWebhook<TPayload> {
+  event: SyncWebhookEventType;
+
+  type = "sync" as const;
+
+  constructor(configuration: WebhookConfig<SyncWebhookEventType>) {
+    super(configuration);
+
+    this.event = configuration.event;
+  }
+}
