+++ src/handlers/next/saleor-webhooks/sync-webhook-response-builder.ts
@@ -1,0 +1,9 @@
+import { SyncWebhookEventType } from "../../../types";
+
+const SyncResponsesMap = {
+  "CHECKOUT_CALCULATE_TAXES": {
+    foo :"bar"
+  }
+} satisfies Partial<Record<SyncWebhookEventType, object>> // todo implement all fields
+
+type SyncResponseBuilder = <E extends "CHECKOUT_CALCULATE_TAXES">() => typeof SyncResponsesMap[E]
